src/requirements.txt
openai
discord.py
requests
beautifulsoup4
